r/aws_sqs_queue: Migrate to AWS SDK for Go v2.
