+++ internal/service/sqs/queue.go
@@ -8,13 +8,17 @@ import (
 	"fmt"
 	"log"
 	"regexp"
+	"strconv"
+	"time"
 
 	"github.com/YakDriver/regexache"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/sqs"
-	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/sqs"
+	"github.com/aws/aws-sdk-go-v2/service/sqs/types"
+	awspolicy "github.com/hashicorp/awspolicyequivalence"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/structure"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
@@ -22,6 +26,7 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/create"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs"
+	tfmaps "github.com/hashicorp/terraform-provider-aws/internal/maps"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
 	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 	"github.com/hashicorp/terraform-provider-aws/internal/verify"
@@ -157,29 +162,29 @@ var (
 		},
 	}
 
-	queueAttributeMap = attrmap.New(map[string]string{
-		"arn":                               sqs.QueueAttributeNameQueueArn,
-		"content_based_deduplication":       sqs.QueueAttributeNameContentBasedDeduplication,
-		"deduplication_scope":               sqs.QueueAttributeNameDeduplicationScope,
-		"delay_seconds":                     sqs.QueueAttributeNameDelaySeconds,
-		"fifo_queue":                        sqs.QueueAttributeNameFifoQueue,
-		"fifo_throughput_limit":             sqs.QueueAttributeNameFifoThroughputLimit,
-		"kms_data_key_reuse_period_seconds": sqs.QueueAttributeNameKmsDataKeyReusePeriodSeconds,
-		"kms_master_key_id":                 sqs.QueueAttributeNameKmsMasterKeyId,
-		"max_message_size":                  sqs.QueueAttributeNameMaximumMessageSize,
-		"message_retention_seconds":         sqs.QueueAttributeNameMessageRetentionPeriod,
-		"policy":                            sqs.QueueAttributeNamePolicy,
-		"receive_wait_time_seconds":         sqs.QueueAttributeNameReceiveMessageWaitTimeSeconds,
-		"redrive_allow_policy":              sqs.QueueAttributeNameRedriveAllowPolicy,
-		"redrive_policy":                    sqs.QueueAttributeNameRedrivePolicy,
-		"sqs_managed_sse_enabled":           sqs.QueueAttributeNameSqsManagedSseEnabled,
-		"visibility_timeout_seconds":        sqs.QueueAttributeNameVisibilityTimeout,
+	queueAttributeMap = attrmap.New(map[string]types.QueueAttributeName{
+		"arn":                               types.QueueAttributeNameQueueArn,
+		"content_based_deduplication":       types.QueueAttributeNameContentBasedDeduplication,
+		"deduplication_scope":               types.QueueAttributeNameDeduplicationScope,
+		"delay_seconds":                     types.QueueAttributeNameDelaySeconds,
+		"fifo_queue":                        types.QueueAttributeNameFifoQueue,
+		"fifo_throughput_limit":             types.QueueAttributeNameFifoThroughputLimit,
+		"kms_data_key_reuse_period_seconds": types.QueueAttributeNameKmsDataKeyReusePeriodSeconds,
+		"kms_master_key_id":                 types.QueueAttributeNameKmsMasterKeyId,
+		"max_message_size":                  types.QueueAttributeNameMaximumMessageSize,
+		"message_retention_seconds":         types.QueueAttributeNameMessageRetentionPeriod,
+		"policy":                            types.QueueAttributeNamePolicy,
+		"receive_wait_time_seconds":         types.QueueAttributeNameReceiveMessageWaitTimeSeconds,
+		"redrive_allow_policy":              types.QueueAttributeNameRedriveAllowPolicy,
+		"redrive_policy":                    types.QueueAttributeNameRedrivePolicy,
+		"sqs_managed_sse_enabled":           types.QueueAttributeNameSqsManagedSseEnabled,
+		"visibility_timeout_seconds":        types.QueueAttributeNameVisibilityTimeout,
 	}, queueSchema).WithIAMPolicyAttribute("policy").WithMissingSetToNil("*").WithAlwaysSendConfiguredBooleanValueOnCreate("sqs_managed_sse_enabled")
 )
 
 // @SDKResource("aws_sqs_queue", name="Queue")
 // @Tags(identifierAttribute="id")
-func ResourceQueue() *schema.Resource {
+func resourceQueue() *schema.Resource {
 	return &schema.Resource{
 		CreateWithoutTimeout: resourceQueueCreate,
 		ReadWithoutTimeout:   resourceQueueRead,
@@ -200,7 +205,7 @@ func ResourceQueue() *schema.Resource {
 }
 
 func resourceQueueCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	conn := meta.(*conns.AWSClient).SQSConn(ctx)
+	conn := meta.(*conns.AWSClient).SQSClient(ctx)
 
 	name := queueName(d)
 	input := &sqs.CreateQueueInput{
@@ -213,26 +218,28 @@ func resourceQueueCreate(ctx context.Context, d *schema.ResourceData, meta inter
 		return diag.FromErr(err)
 	}
 
-	input.Attributes = aws.StringMap(attributes)
+	input.Attributes = tfmaps.ApplyToAllKeys(attributes, func(v types.QueueAttributeName) string {
+		return string(v)
+	})
 
-	outputRaw, err := tfresource.RetryWhenAWSErrCodeEquals(ctx, queueCreatedTimeout, func() (interface{}, error) {
-		return conn.CreateQueueWithContext(ctx, input)
-	}, sqs.ErrCodeQueueDeletedRecently)
+	outputRaw, err := tfresource.RetryWhenIsA[*types.QueueDeletedRecently](ctx, queueCreatedTimeout, func() (interface{}, error) {
+		return conn.CreateQueue(ctx, input)
+	})
 
 	// Some partitions (e.g. ISO) may not support tag-on-create.
-	if input.Tags != nil && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
+	if input.Tags != nil && errs.IsUnsupportedOperationInPartitionError(meta.(*conns.AWSClient).Partition, err) {
 		input.Tags = nil
 
-		outputRaw, err = tfresource.RetryWhenAWSErrCodeEquals(ctx, queueCreatedTimeout, func() (interface{}, error) {
-			return conn.CreateQueueWithContext(ctx, input)
-		}, sqs.ErrCodeQueueDeletedRecently)
+		outputRaw, err = tfresource.RetryWhenIsA[*types.QueueDeletedRecently](ctx, queueCreatedTimeout, func() (interface{}, error) {
+			return conn.CreateQueue(ctx, input)
+		})
 	}
 
 	if err != nil {
 		return diag.Errorf("creating SQS Queue (%s): %s", name, err)
 	}
 
-	d.SetId(aws.StringValue(outputRaw.(*sqs.CreateQueueOutput).QueueUrl))
+	d.SetId(aws.ToString(outputRaw.(*sqs.CreateQueueOutput).QueueUrl))
 
 	if err := waitQueueAttributesPropagated(ctx, conn, d.Id(), attributes); err != nil {
 		return diag.Errorf("waiting for SQS Queue (%s) attributes create: %s", d.Id(), err)
@@ -243,7 +250,7 @@ func resourceQueueCreate(ctx context.Context, d *schema.ResourceData, meta inter
 		err := createTags(ctx, conn, d.Id(), tags)
 
 		// If default tags only, continue. Otherwise, error.
-		if v, ok := d.GetOk(names.AttrTags); (!ok || len(v.(map[string]interface{})) == 0) && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
+		if v, ok := d.GetOk(names.AttrTags); (!ok || len(v.(map[string]interface{})) == 0) && errs.IsUnsupportedOperationInPartitionError(meta.(*conns.AWSClient).Partition, err) {
 			return resourceQueueRead(ctx, d, meta)
 		}
 
@@ -256,10 +263,10 @@ func resourceQueueCreate(ctx context.Context, d *schema.ResourceData, meta inter
 }
 
 func resourceQueueRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	conn := meta.(*conns.AWSClient).SQSConn(ctx)
+	conn := meta.(*conns.AWSClient).SQSClient(ctx)
 
 	outputRaw, err := tfresource.RetryWhenNotFound(ctx, queueReadTimeout, func() (interface{}, error) {
-		return FindQueueAttributesByURL(ctx, conn, d.Id())
+		return findQueueAttributesByURL(ctx, conn, d.Id())
 	})
 
 	if !d.IsNewResource() && tfresource.NotFound(err) {
@@ -277,7 +284,7 @@ func resourceQueueRead(ctx context.Context, d *schema.ResourceData, meta interfa
 		return diag.FromErr(err)
 	}
 
-	err = queueAttributeMap.APIAttributesToResourceData(outputRaw.(map[string]string), d)
+	err = queueAttributeMap.APIAttributesToResourceData(outputRaw.(map[types.QueueAttributeName]string), d)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -299,7 +306,7 @@ func resourceQueueRead(ctx context.Context, d *schema.ResourceData, meta interfa
 }
 
 func resourceQueueUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	conn := meta.(*conns.AWSClient).SQSConn(ctx)
+	conn := meta.(*conns.AWSClient).SQSClient(ctx)
 
 	if d.HasChangesExcept("tags", "tags_all") {
 		attributes, err := queueAttributeMap.ResourceDataToAPIAttributesUpdate(d)
@@ -308,11 +315,14 @@ func resourceQueueUpdate(ctx context.Context, d *schema.ResourceData, meta inter
 		}
 
 		input := &sqs.SetQueueAttributesInput{
-			Attributes: aws.StringMap(attributes),
-			QueueUrl:   aws.String(d.Id()),
+			QueueUrl: aws.String(d.Id()),
 		}
 
-		_, err = conn.SetQueueAttributesWithContext(ctx, input)
+		input.Attributes = tfmaps.ApplyToAllKeys(attributes, func(v types.QueueAttributeName) string {
+			return string(v)
+		})
+
+		_, err = conn.SetQueueAttributes(ctx, input)
 
 		if err != nil {
 			return diag.Errorf("updating SQS Queue (%s) attributes: %s", d.Id(), err)
@@ -327,14 +337,14 @@ func resourceQueueUpdate(ctx context.Context, d *schema.ResourceData, meta inter
 }
 
 func resourceQueueDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	conn := meta.(*conns.AWSClient).SQSConn(ctx)
+	conn := meta.(*conns.AWSClient).SQSClient(ctx)
 
 	log.Printf("[DEBUG] Deleting SQS Queue: %s", d.Id())
-	_, err := conn.DeleteQueueWithContext(ctx, &sqs.DeleteQueueInput{
+	_, err := conn.DeleteQueue(ctx, &sqs.DeleteQueueInput{
 		QueueUrl: aws.String(d.Id()),
 	})
 
-	if tfawserr.ErrCodeEquals(err, sqs.ErrCodeQueueDoesNotExist) {
+	if errs.IsA[*types.QueueDoesNotExist](err) {
 		return nil
 	}
 
@@ -382,4 +392,193 @@ func queueName(d interface{ Get(string) any }) string {
 		optFns = append(optFns, create.WithSuffix(FIFOQueueNameSuffix))
 	}
 	return create.NewNameGenerator(optFns...).Generate()
+}
+
+func findQueueAttributesByURL(ctx context.Context, conn *sqs.Client, url string) (map[types.QueueAttributeName]string, error) {
+	input := &sqs.GetQueueAttributesInput{
+		AttributeNames: []types.QueueAttributeName{types.QueueAttributeNameAll},
+		QueueUrl:       aws.String(url),
+	}
+
+	output, err := findQueueAttributes(ctx, conn, input)
+
+	if err != nil {
+		return nil, err
+	}
+
+	return output, nil
+}
+
+func findQueueAttributeByTwoPartKey(ctx context.Context, conn *sqs.Client, url string, attributeName types.QueueAttributeName) (*string, error) {
+	input := &sqs.GetQueueAttributesInput{
+		AttributeNames: []types.QueueAttributeName{attributeName},
+		QueueUrl:       aws.String(url),
+	}
+
+	output, err := findQueueAttributes(ctx, conn, input)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if v, ok := output[attributeName]; ok && v != "" {
+		return &v, nil
+	}
+
+	return nil, tfresource.NewEmptyResultError(input)
+}
+
+func findQueueAttributes(ctx context.Context, conn *sqs.Client, input *sqs.GetQueueAttributesInput) (map[types.QueueAttributeName]string, error) {
+	output, err := conn.GetQueueAttributes(ctx, input)
+
+	if errs.IsA[*types.QueueDoesNotExist](err) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || len(output.Attributes) == 0 {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return tfmaps.ApplyToAllKeys(output.Attributes, func(v string) types.QueueAttributeName {
+		return types.QueueAttributeName(v)
+	}), nil
+}
+
+const (
+	// Maximum amount of time to wait for SQS queue attribute changes to propagate
+	// This timeout should not be increased without strong consideration
+	// as this will negatively impact user experience when configurations
+	// have incorrect references or permissions.
+	// Reference: https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SetQueueAttributes.html
+	queueAttributePropagationTimeout = 2 * time.Minute
+
+	// If you delete a queue, you must wait at least 60 seconds before creating a queue with the same name.
+	// ReferenceL https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_CreateQueue.html
+	queueCreatedTimeout = 70 * time.Second
+	queueReadTimeout    = 20 * time.Second
+	queueDeletedTimeout = 3 * time.Minute
+	queueTagsTimeout    = 60 * time.Second
+
+	queueAttributeReadTimeout = 20 * time.Second
+
+	queueStateExists = "exists"
+
+	queueAttributeStateNotEqual = "notequal"
+	queueAttributeStateEqual    = "equal"
+)
+
+func statusQueueState(ctx context.Context, conn *sqs.Client, url string) retry.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		output, err := findQueueAttributesByURL(ctx, conn, url)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, queueStateExists, nil
+	}
+}
+
+func statusQueueAttributeState(ctx context.Context, conn *sqs.Client, url string, expected map[types.QueueAttributeName]string) retry.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		attributesMatch := func(got map[types.QueueAttributeName]string) string {
+			for k, e := range expected {
+				g, ok := got[k]
+
+				if !ok {
+					// Missing attribute equivalent to empty expected value.
+					if e == "" {
+						continue
+					}
+
+					// Backwards compatibility: https://github.com/hashicorp/terraform-provider-aws/issues/19786.
+					if k == types.QueueAttributeNameKmsDataKeyReusePeriodSeconds && e == strconv.Itoa(DefaultQueueKMSDataKeyReusePeriodSeconds) {
+						continue
+					}
+
+					return queueAttributeStateNotEqual
+				}
+
+				switch k {
+				case types.QueueAttributeNamePolicy:
+					equivalent, err := awspolicy.PoliciesAreEquivalent(g, e)
+
+					if err != nil {
+						return queueAttributeStateNotEqual
+					}
+
+					if !equivalent {
+						return queueAttributeStateNotEqual
+					}
+				case types.QueueAttributeNameRedriveAllowPolicy, types.QueueAttributeNameRedrivePolicy:
+					if !StringsEquivalent(g, e) {
+						return queueAttributeStateNotEqual
+					}
+				default:
+					if g != e {
+						return queueAttributeStateNotEqual
+					}
+				}
+			}
+
+			return queueAttributeStateEqual
+		}
+
+		got, err := findQueueAttributesByURL(ctx, conn, url)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		status := attributesMatch(got)
+
+		return got, status, nil
+	}
+}
+
+func waitQueueAttributesPropagated(ctx context.Context, conn *sqs.Client, url string, expected map[types.QueueAttributeName]string) error {
+	stateConf := &retry.StateChangeConf{
+		Pending:                   []string{queueAttributeStateNotEqual},
+		Target:                    []string{queueAttributeStateEqual},
+		Refresh:                   statusQueueAttributeState(ctx, conn, url, expected),
+		Timeout:                   queueAttributePropagationTimeout,
+		ContinuousTargetOccurence: 6,               // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+		MinTimeout:                5 * time.Second, // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+		NotFoundChecks:            10,              // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+	}
+
+	_, err := stateConf.WaitForStateContext(ctx)
+
+	return err
+}
+
+func waitQueueDeleted(ctx context.Context, conn *sqs.Client, url string) error {
+	stateConf := &retry.StateChangeConf{
+		Pending:                   []string{queueStateExists},
+		Target:                    []string{},
+		Refresh:                   statusQueueState(ctx, conn, url),
+		Timeout:                   queueDeletedTimeout,
+		ContinuousTargetOccurence: 15,              // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+		MinTimeout:                3 * time.Second, // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+		NotFoundChecks:            5,               // set to accommodate GovCloud, commercial, China, etc. - avoid lowering
+	}
+
+	_, err := stateConf.WaitForStateContext(ctx)
+
+	return err
 }
