+++ internal/service/sqs/service_package_gen.go
@@ -38,7 +38,7 @@ func (p *servicePackage) SDKDataSources(ctx context.Context) []*types.ServicePac
 func (p *servicePackage) SDKResources(ctx context.Context) []*types.ServicePackageSDKResource {
 	return []*types.ServicePackageSDKResource{
 		{
-			Factory:  ResourceQueue,
+			Factory:  resourceQueue,
 			TypeName: "aws_sqs_queue",
 			Name:     "Queue",
 			Tags: &types.ServicePackageResourceTags{
